Coverage SQL: Closed a window of vulnerability (race condition). When invoking synchronized(getLock()), the Database.Session.monitorExit(...) could be invoked in a background thread during the slight time window between the return of Table.getLock() and the synchronized statement, resulting in the Database.Session lock be declared as available for an other thread (its threadID field is set to null), or disposed. When disposed, the call first call to any method requirying this session (for example 'transactionBegin') result in the creation of a new Database.Session object, which leads to the assertion failure seen in assert(Thread.holdsLock(...)) statements.
The fix is to pass the LocalCache argument explicitly to any method needing it, so making sure that
we really use the same instance. It doesn't solve the race condition (so this fix is not completly
satisfying), but it ensure that the appropriate synchronization locks are used when the above-cited
race condition happen. In this worst case scenario, we have thread contention but at least the
computational behavior should be correct.