test(Storage): temporarily ignore a test which is caused by a slight difference between the coverage CRS and the expected CRS.
Investigation has shown a difference in the 3 last digits of a coefficients of an affine transform (about 0.0001 millimetres).
We needs more investigation in order to understand where this difference come from. For now we suspect a difference in the accuracy
of SIS calculation compared to Geotk calculation. The comparison has a tolerance threshold, but this threshold is deep inside SIS
and we prefer to wait for the completion of SIS port before the relax the threshold (we can easily relax threshold later, but the
converse is much more difficult).