refactor(*): avoid usage of deprecated SIS 0.6 methods to be removed in SIS 0.7.
The vast majority of changes was to replace calls to Logging.getLogger(Class) by calls to Logging.getLogger(String).
The method expecting a Class argument has been deprecated because a look to how this method was used has shown that
it was causing more hurt than good. For example there is many duplicated instantiations of a static final LOGGER field,
maybe the fact that getLogger(Class) looks only at the package name (ignoring the class name) was not obvious.
Conversely, they were many calls to Logger.getLogger(String) with 'SomeClass.getName()' in argument, maybe because
some peoples followed the Logging.getLogger(Class) pattern without realizing that only the package name was supposed
to be used.